+++ src/ILInspector.Decompiler/Pipeline/Ir/CSharpPrinter.RaisedExpressions.cs
@@ -1,0 +1,158 @@
+using System.Text;
+
+namespace ILInspector.Decompiler.Pipeline;
+
+/// <summary>Rendering helpers for raised high-level expression nodes.</summary>
+public sealed partial class CSharpPrinter
+{
+    /// <summary>
+    /// Renders a raised object/collection initializer: <c>new T(args) { ... }</c>
+    /// where the body is <c>Member = value</c> entries (object form) or bare
+    /// element expressions (collection form). Constructor parens are omitted when
+    /// the creation takes no arguments, matching idiomatic C#.
+    /// </summary>
+    string ObjectInitializerText(ObjectInitializerExpression initializer)
+    {
+        var creation = initializer.Creation;
+        var arguments = creation.Arguments.Count == 0
+            ? string.Empty
+            : $"({Arguments(creation.Arguments, creation.Constructor.ParameterTypes, creation.Constructor.ParameterRefKinds)})";
+        var body = initializer.IsCollection
+            ? string.Join(", ", initializer.Values.Select(Expression))
+            : string.Join(", ", initializer.Members.Zip(initializer.Values, (member, value) => $"{member} = {Expression(value)}"));
+        return $"new {TypeText(creation.Constructor.DeclaringType)}{arguments} {{ {body} }}";
+    }
+
+    /// <summary>
+    /// Renders <c>new { Name = value, ... }</c>. A member uses the projection
+    /// shorthand (<c>new { x }</c> / <c>new { obj.Member }</c>) only when the value
+    /// expression's own member name exactly matches the property name — an
+    /// identifier whose text equals the name, or a field/property access ending in
+    /// <c>.Name</c>. Any mismatch keeps the explicit <c>Name = value</c> form, so
+    /// the recovered name is never silently changed.
+    /// </summary>
+    string AnonymousObjectText(AnonymousObject anonymous)
+    {
+        if (anonymous.Values.Count == 0)
+            return "new { }";
+        var parts = new List<string>(anonymous.Values.Count);
+        for (int i = 0; i < anonymous.Values.Count; i++)
+        {
+            var value = anonymous.Values[i];
+            string name = anonymous.PropertyNames[i];
+            string text = Expression(value);
+            bool shorthand = text == name
+                || (value is LoadField field && field.Field.Name == name && text.EndsWith("." + name, StringComparison.Ordinal))
+                || (value is LoadProperty property && property.PropertyName == name && text.EndsWith("." + name, StringComparison.Ordinal));
+            parts.Add(shorthand ? text : $"{name} = {text}");
+        }
+        return $"new {{ {string.Join(", ", parts)} }}";
+    }
+
+    /// <summary>
+    /// Renders a recovered lambda: parameter names (the delegate type supplies
+    /// their types, so they stay unannotated), then <c>=&gt; expr</c> for an
+    /// expression body or <c>=&gt; { ... }</c> otherwise. A single parameter
+    /// drops the parentheses. The body's arguments are self-naming on their
+    /// nodes, so this reuses the outer printer with no scope switch — sound only
+    /// because the raising pass admits only non-capturing, zero-local bodies.
+    /// </summary>
+    string LambdaText(Lambda lambda)
+    {
+        string parameters = lambda.Parameters is [var single]
+            ? single.Name
+            : $"({string.Join(", ", lambda.Parameters.Select(p => p.Name))})";
+
+        if (lambda.ExpressionBody is { } expr)
+            return $"{parameters} => {Expression(expr)}";
+
+        var statements = lambda.Body.Blocks
+            .SelectMany(b => b.Children)
+            .Select(Statement)
+            .Where(s => s is not null);
+        return $"{parameters} => {{ {string.Join(" ", statements)} }}";
+    }
+
+    /// <summary>The text of one switch-expression arm: its labels (or <c>_</c>) and the value it yields.</summary>
+    string SwitchArmText(SwitchExpressionArm arm)
+        => $"{(arm.IsDefault ? "_" : string.Join(" or ", arm.Labels))} => {Expression(arm.Value)}";
+
+    /// <summary>The single-line form of a switch expression, used when it is nested inside another expression.</summary>
+    string SwitchExpressionInline(SwitchExpression node)
+        => $"{Operand(node.Value)} switch {{ {string.Join(", ", node.Arms.Select(SwitchArmText))} }}";
+
+    string InterpolatedStringText(InterpolatedStringExpression node)
+    {
+        var sb = new StringBuilder().Append("$\"");
+        foreach (var part in node.Parts)
+        {
+            if (part.IsLiteral)
+            {
+                sb.Append(InterpolatedLiteralText(part.Literal!));
+            }
+            else if (part.ExpressionIndex >= 0 && part.ExpressionIndex < node.FormattedValues.Count)
+            {
+                sb.Append('{').Append(InterpolatedExpression(node.FormattedValues[part.ExpressionIndex])).Append('}');
+            }
+        }
+        return sb.Append('"').ToString();
+    }
+
+    string InterpolatedExpression(IrExpression value)
+        => value is LoadArgument or LoadLocal or LoadStackSlot or Constant or LoadField or LoadProperty or Call
+            ? Expression(value)
+            : $"({Expression(value)})";
+
+    static string InterpolatedLiteralText(string value)
+    {
+        var sb = new StringBuilder(value.Length);
+        foreach (char c in value)
+        {
+            if (c == '{')
+                sb.Append("{{");
+            else if (c == '}')
+                sb.Append("}}");
+            else
+                sb.Append(EscapeChar(c, inString: true));
+        }
+        return sb.ToString();
+    }
+
+    /// <summary>
+    /// <c>target?.Member</c>: the member's receiver child is the target, and the
+    /// member's name/arguments form the suffix after <c>?</c>. Mirrors the
+    /// instance spellings of <see cref="CallText"/>, <see cref="PropertyTarget"/>,
+    /// and <see cref="FieldTarget"/>, minus their receiver — the <c>?.</c> owns it.
+    /// </summary>
+    string NullConditionalText(NullConditional node)
+    {
+        var member = node.Member;
+        var receiver = NullConditionalReceiver(member);
+        return $"{ReceiverText(receiver)}?{NullConditionalSuffix(member)}";
+    }
+
+    static IrExpression NullConditionalReceiver(IrExpression member) => member switch
+    {
+        Call call => call.Arguments[0],
+        LoadProperty property => property.Instance!,
+        LoadField field => field.Instance!,
+        _ => member,
+    };
+
+    string NullConditionalSuffix(IrExpression member) => member switch
+    {
+        LoadField field => $".{field.Field.Name}",
+        LoadProperty property when property.IndexArguments.Count > 0 => $"[{Arguments(property.IndexArguments)}]",
+        LoadProperty property => $".{property.PropertyName}",
+        Call call => NullConditionalCallSuffix(call),
+        _ => $".{member.Describe()}",
+    };
+
+    string NullConditionalCallSuffix(Call call)
+    {
+        string typeArguments = call.Callee.TypeArguments.IsEmpty
+            ? ""
+            : $"<{string.Join(", ", call.Callee.TypeArguments.Select(TypeText))}>";
+        return $".{CSharpNaming.MethodName(call.Callee.Name)}{typeArguments}({Arguments(call.Arguments.Skip(1), call.Callee.ParameterTypes, call.Callee.ParameterRefKinds)})";
+    }
+}
